refactor(ecosystem-automation/golang): strip Weaver registry generation into feat/weaver-registry
Remove all Weaver-specific code (generator, Group/AttributeRef/AttributeDef types,
CalculateStats, convertTelemetryToGroups) so this branch is solely about walking
go-contrib and emitting the versioned inventory. Weaver registry output lives on
feat/weaver-registry branched from this HEAD.
