+++ ecosystem-automation/golang-instrumentation-watcher/.gitignore
@@ -25,10 +25,6 @@ go.work.sum
 opentelemetry-go*
 semantic-conventions*
 
-# build
-# instrumentation-list.yaml
-registry/
-
 # llms
 CLAUDE.md
 CLAUDE.local.md
