--- ecosystem-automation/golang-instrumentation-watcher/cmd/snapshot/main.go
@@ -1,274 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"crypto/sha256"
-	"fmt"
-	"os"
-	"path/filepath"
-	"slices"
-	"strconv"
-	"strings"
-
-	"gopkg.in/yaml.v3"
-
-	"github.com/open-telemetry/opentelemetry-ecosystem-explorer/golang-instrumentation-watcher/instrumentation"
-	"github.com/open-telemetry/opentelemetry-ecosystem-explorer/golang-instrumentation-watcher/inventory"
-	"github.com/open-telemetry/opentelemetry-ecosystem-explorer/golang-instrumentation-watcher/metadata"
-)
-
-const sentinel = "<!-- otel-go-snapshot -->"
-
-func main() {
-	if err := run(); err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-	}
-}
-
-func run() error {
-	workDir, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-	root, err := findRepoRoot(workDir)
-	if err != nil {
-		return err
-	}
-
-	inventoryDir := filepath.Join(root, "ecosystem-registry", "go", "contrib")
-	mgr := inventory.NewManager(inventoryDir)
-
-	releases, err := mgr.ListReleaseVersions()
-	if err != nil {
-		return err
-	}
-	if len(releases) == 0 {
-		return fmt.Errorf("no release versions found; run make sync first")
-	}
-	releaseVer := releases[0]
-
-	inv, err := mgr.Load(releaseVer)
-	if err != nil {
-		return err
-	}
-
-	releaseData, err := os.ReadFile(filepath.Join(mgr.VersionDir(releaseVer), "instrumentation.yaml"))
-	if err != nil {
-		return err
-	}
-	releaseLines := bytes.Count(releaseData, []byte("\n"))
-	releaseHash := fileHash(releaseData)
-
-	snapshotVer, snapshotHash := "", ""
-	if snapshots, _ := mgr.ListSnapshotVersions(); len(snapshots) > 0 {
-		snapshotVer = snapshots[0]
-		if snapData, err := os.ReadFile(filepath.Join(mgr.VersionDir(snapshotVer), "instrumentation.yaml")); err == nil {
-			snapshotHash = fileHash(snapData)
-		}
-	}
-
-	fmt.Print(render(inv.Libraries, releaseVer, releaseHash, releaseLines, snapshotVer, snapshotHash))
-	return nil
-}
-
-func render(libs []instrumentation.Library, releaseVer, releaseHash string, releaseLines int, snapshotVer, snapshotHash string) string {
-	var b strings.Builder
-
-	fmt.Fprintln(&b, sentinel)
-	fmt.Fprintln(&b, "## 📦 Rendered output preview (for reviewers)")
-	fmt.Fprintln(&b)
-	fmt.Fprintln(&b, "The inventory files themselves are not in this PR's diff — they're produced by the nightly CI automation (`ecosystem-registry/go/` is generated, like the other watchers). This comment shows what `make sync` actually emits so reviewers can evaluate the output shape without running it.")
-	fmt.Fprintln(&b)
-	fmt.Fprintln(&b, "Generated from a local `make sync` against **opentelemetry-go-contrib**:")
-	fmt.Fprintf(&b, "- Release: **`%s`** → `ecosystem-registry/go/contrib/%s/instrumentation.yaml` (sha256/12 `%s`)\n", releaseVer, releaseVer, releaseHash)
-	if snapshotVer != "" {
-		fmt.Fprintf(&b, "- Snapshot: **`%s`** from `main` (sha256/12 `%s`)\n", snapshotVer, snapshotHash)
-	}
-	fmt.Fprintf(&b, "- **%d libraries**, %d lines each. Byte-identical across repeated runs (determinism verified).\n", len(libs), releaseLines)
-	fmt.Fprintln(&b)
-
-	sorted := sortedLibs(libs)
-	nameCount := countDisplayNames(libs)
-
-	fmt.Fprintf(&b, "### Libraries (%d)\n", len(libs))
-	fmt.Fprintln(&b)
-	fmt.Fprintln(&b, "Scope is components that **instrument a developer's code**: instrumentation wrappers + log bridges. Pipeline-config components (exporters, propagators, samplers, detectors, processors) are intentionally excluded — see PR description.")
-	fmt.Fprintln(&b)
-	fmt.Fprintln(&b, "| display_name | type | module version |")
-	fmt.Fprintln(&b, "|---|---|---|")
-	for _, lib := range sorted {
-		ver := lib.Module.Version
-		if ver == "" {
-			ver = "_(no release tag yet)_"
-		}
-		fmt.Fprintf(&b, "| %s | %s | %s |\n", displayLabel(lib, nameCount), lib.InstrumentationType, ver)
-	}
-	fmt.Fprintln(&b)
-
-	if hasDuplicateNames(nameCount) {
-		fmt.Fprintln(&b, "> **Note for reviewers:** some libraries share a display name (e.g. MongoDB v1 and v2 drivers). The `name` field is derived from the full module path to keep records unique. A driver showing an empty `version` has no per-module release tag yet — it picks one up once go-contrib tags it.")
-		fmt.Fprintln(&b)
-	}
-
-	for _, lib := range pickFeatured(sorted) {
-		fmt.Fprintf(&b, "### Example record — `%s` (%s)\n", shortName(lib), featuredDesc(lib))
-		fmt.Fprintln(&b)
-		fmt.Fprintln(&b, "```yaml")
-		fmt.Fprint(&b, asListItem(lib))
-		fmt.Fprintln(&b, "```")
-		fmt.Fprintln(&b)
-	}
-
-	fmt.Fprintln(&b, "_Generated by `make snapshot`; the nightly workflow integration is tracked separately._")
-	return b.String()
-}
-
-func sortedLibs(libs []instrumentation.Library) []instrumentation.Library {
-	out := make([]instrumentation.Library, len(libs))
-	copy(out, libs)
-	slices.SortFunc(out, func(a, b instrumentation.Library) int {
-		ta, tb := a.InstrumentationType.String(), b.InstrumentationType.String()
-		if ta != tb {
-			return strings.Compare(ta, tb) // bridge < wrapper
-		}
-		return strings.Compare(strings.ToLower(a.DisplayName), strings.ToLower(b.DisplayName))
-	})
-	return out
-}
-
-func countDisplayNames(libs []instrumentation.Library) map[string]int {
-	m := make(map[string]int, len(libs))
-	for _, lib := range libs {
-		m[lib.DisplayName]++
-	}
-	return m
-}
-
-func hasDuplicateNames(nameCount map[string]int) bool {
-	for _, n := range nameCount {
-		if n > 1 {
-			return true
-		}
-	}
-	return false
-}
-
-func displayLabel(lib instrumentation.Library, nameCount map[string]int) string {
-	if nameCount[lib.DisplayName] <= 1 {
-		return lib.DisplayName
-	}
-	for _, seg := range strings.Split(lib.Module.Path, "/") {
-		if len(seg) > 1 && seg[0] == 'v' {
-			if _, err := strconv.Atoi(seg[1:]); err == nil {
-				return lib.DisplayName + " (" + seg + ")"
-			}
-		}
-	}
-	return lib.DisplayName + " (v1)"
-}
-
-func pickFeatured(sorted []instrumentation.Library) []instrumentation.Library {
-	var bestWrapper *instrumentation.Library
-	bestScore := -1
-	var firstBridge *instrumentation.Library
-
-	for i := range sorted {
-		lib := &sorted[i]
-		switch lib.InstrumentationType {
-		case metadata.InstrTypeWrapper:
-			if s := telemetryScore(*lib); s > bestScore {
-				bestScore = s
-				bestWrapper = lib
-			}
-		case metadata.InstrTypeBridge:
-			if firstBridge == nil {
-				firstBridge = lib
-			}
-		}
-	}
-
-	var out []instrumentation.Library
-	if bestWrapper != nil {
-		out = append(out, *bestWrapper)
-	}
-	if firstBridge != nil {
-		out = append(out, *firstBridge)
-	}
-	return out
-}
-
-func telemetryScore(lib instrumentation.Library) int {
-	score := 0
-	for _, tel := range lib.Telemetry {
-		score += len(tel.Spans) + len(tel.Metrics)
-		for _, span := range tel.Spans {
-			score += len(span.Attributes)
-		}
-		for _, metric := range tel.Metrics {
-			score += len(metric.Attributes)
-		}
-	}
-	return score
-}
-
-func shortName(lib instrumentation.Library) string {
-	parts := strings.Split(lib.Name, "-")
-	return parts[len(parts)-1]
-}
-
-func featuredDesc(lib instrumentation.Library) string {
-	switch lib.InstrumentationType {
-	case metadata.InstrTypeBridge:
-		return "log bridge"
-	case metadata.InstrTypeWrapper:
-		for _, tel := range lib.Telemetry {
-			if len(tel.Spans) > 0 && len(tel.Metrics) > 0 {
-				return "spans + metrics fused"
-			}
-		}
-		return "instrumentation wrapper"
-	default:
-		return lib.InstrumentationType.String()
-	}
-}
-
-func asListItem(lib instrumentation.Library) string {
-	var buf bytes.Buffer
-	enc := yaml.NewEncoder(&buf)
-	enc.SetIndent(2)
-	_ = enc.Encode(lib)
-	_ = enc.Close()
-
-	lines := strings.Split(strings.TrimRight(buf.String(), "\n"), "\n")
-	var b strings.Builder
-	for i, line := range lines {
-		if i == 0 {
-			fmt.Fprintln(&b, "- "+line)
-		} else if line == "" {
-			fmt.Fprintln(&b, "")
-		} else {
-			fmt.Fprintln(&b, "  "+line)
-		}
-	}
-	return b.String()
-}
-
-func fileHash(data []byte) string {
-	sum := sha256.Sum256(data)
-	return fmt.Sprintf("%x", sum)[:12]
-}
-
-func findRepoRoot(dir string) (string, error) {
-	current := dir
-	for {
-		if _, err := os.Stat(filepath.Join(current, "ecosystem-registry")); err == nil {
-			return current, nil
-		}
-		parent := filepath.Dir(current)
-		if parent == current {
-			return "", fmt.Errorf("could not locate repo root (no ecosystem-registry/ ancestor of %s)", dir)
-		}
-		current = parent
-	}
-}
