+++ ecosystem-automation/golang-instrumentation-watcher/instrumentation/instrumentation_test.go
@@ -197,80 +197,6 @@ func TestLambdaInstrumentation(t *testing.T) {
 }
 
 func TestFullScanValidation(t *testing.T) {
-	t.Run("scan all instrumentation packages", func(t *testing.T) {
-		repoPath := getRepoPath(t)
-		result, err := ScanRepo(repo.RepoContrib, repoPath)
-		if err != nil {
-			t.Fatalf("ScanRepo() error = %v", err)
-		}
-		groups := result.Groups
-
-		if got := len(groups); got < 10 {
-			t.Errorf("Total groups = %d, want at least 10", got)
-		}
-
-		groupsByRepo := map[string][]Group{
-			repo.RepoContrib: groups,
-		}
-		repoStats := CalculateStats(groupsByRepo)
-		stats := repoStats[repo.RepoContrib]
-
-		if stats.TotalSpans < 10 {
-			t.Errorf("Total spans = %d, want at least 10", stats.TotalSpans)
-		}
-
-		if stats.TotalMetrics < 15 {
-			t.Errorf("Total metrics = %d, want at least 15", stats.TotalMetrics)
-		}
-
-		t.Logf("Stats breakdown:")
-		t.Logf("  Total groups: %d", len(groups))
-		t.Logf("  Total spans: %d", stats.TotalSpans)
-		t.Logf("  Total metrics: %d", stats.TotalMetrics)
-		t.Logf("  Total attributes: %d", stats.TotalAttributes)
-	})
-
-	t.Run("validate no duplicate groups", func(t *testing.T) {
-		repoPath := getRepoPath(t)
-		result, err := ScanRepo(repo.RepoContrib, repoPath)
-		if err != nil {
-			t.Fatalf("ScanRepo() error = %v", err)
-		}
-		groups := result.Groups
-
-		groupIDs := make(map[string]int)
-		for _, group := range groups {
-			groupIDs[group.ID]++
-		}
-
-		for id, count := range groupIDs {
-			if count > 1 {
-				t.Errorf("Duplicate group ID %s appears %d times", id, count)
-			}
-		}
-	})
-
-	t.Run("validate groups have required fields", func(t *testing.T) {
-		repoPath := getRepoPath(t)
-		result, err := ScanRepo(repo.RepoContrib, repoPath)
-		if err != nil {
-			t.Fatalf("ScanRepo() error = %v", err)
-		}
-		groups := result.Groups
-
-		for _, group := range groups {
-			if group.ID == "" {
-				t.Errorf("Group has empty ID")
-			}
-			if group.Type == "" {
-				t.Errorf("Group %s has empty type", group.ID)
-			}
-			if group.Type != "span" && group.Type != "metric" {
-				t.Errorf("Group %s has invalid type %s", group.ID, group.Type)
-			}
-		}
-	})
-
 	t.Run("scan covers instrumentation and bridges", func(t *testing.T) {
 		repoPath := getRepoPath(t)
 		result, err := ScanRepo(repo.RepoContrib, repoPath)
